+++ src/test/java/org/scion/jpan/testutil/MockBorderRouter.java
@@ -22,9 +22,7 @@ import java.nio.channels.DatagramChannel;
 import java.nio.channels.SelectionKey;
 import java.nio.channels.Selector;
 import java.util.Iterator;
-import org.scion.jpan.Constants;
 import org.scion.jpan.PackageVisibilityHelper;
-import org.scion.jpan.ScionUtil;
 import org.scion.jpan.Scmp;
 import org.scion.jpan.demo.inspector.ScionPacketInspector;
 import org.scion.jpan.demo.inspector.ScmpHeader;
@@ -125,8 +123,7 @@ public class MockBorderRouter implements Runnable {
   private void forwardPacket(ByteBuffer buffer, SocketAddress srcAddress, DatagramChannel outgoing)
       throws IOException {
     InetSocketAddress dstAddress = PackageVisibilityHelper.getDstAddress(buffer);
-    if (ScionUtil.getPropertyOrEnv(
-        Constants.PROPERTY_SHIM, Constants.ENV_SHIM, Constants.DEFAULT_SHIM)) {
+    if (MockNetwork.useShim()) {
       dstAddress = MockNetwork.asInfo.mapDispatcherPorts(dstAddress);
     }
     logger.info(